[fix] test_ps_flex_fwd_flash_bwd: use bf16 to satisfy FlashAttention dtype constraint
FlashAttention only supports fp16/bf16. Switch default test dtype from
fp32 to bf16 so GPU numerical tests can run.

+++ prefix-sharing/tests/unit_test/test_ps_flex_fwd_flash_bwd.py
@@ -49,7 +49,7 @@ def _make_layout(plan, align_size=1):
     return PackedBatchLayout.from_kept_position_rows(rows, align_size=align_size)
 
 
-def _random_qkv(total, num_heads=4, head_dim=64, dtype=torch.float32, seed=42, requires_grad=False):
+def _random_qkv(total, num_heads=4, head_dim=64, dtype=torch.bfloat16, seed=42, requires_grad=False):
     torch.manual_seed(seed)
     q = torch.randn(total, num_heads, head_dim, dtype=dtype, requires_grad=requires_grad)
     k = torch.randn(total, num_heads, head_dim, dtype=dtype, requires_grad=requires_grad)
